+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+ - Add `team` attribute to `Message` and `Command` models. This stores the ID of the Slack Team that the call originated from.
 ### Changed
  - Allow configuration to be partially updated ([@RealOrangeOne](https://github.com/RealOrangeOne) in [#81](https://github.com/sedders123/phial/pull/81))
 ### Fixed